requirements.txt
boto3==1.34.71
botocore==1.34.71
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
idna==3.6
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
typing==3.7.4.3
urllib3==2.2.1
xmltodict==0.13.0
zope.interface==6.2
